+++ examples/intersection.py
@@ -1,8 +1,6 @@
 import math
 
-from fast_trimesh.cpu.io import save_stl
-from fast_trimesh.cpu.shapes import tetrahedron
-from fast_trimesh.cpu.types import Affine3D
+from fast_trimesh import Affine3D, save_stl, tetrahedron
 
 scale = Affine3D(scale=1.5)
 trans = Affine3D(trans=(0.0, 0.0, -1.5))
