+++ examples/simple.py
@@ -1,8 +1,6 @@
 import math
 
-from fast_trimesh.cpu.io import save_stl
-from fast_trimesh.cpu.shapes import cuboid
-from fast_trimesh.cpu.types import Affine3D
+from fast_trimesh import Affine3D, cuboid, save_stl
 
 # Create some cuboids.
 a = cuboid(1.0, 1.0, 1.0, center=True)
