+++ examples/sphere.py
@@ -1,6 +1,4 @@
-from fast_trimesh.cpu.io import save_stl
-from fast_trimesh.cpu.shapes import cuboid, icosphere, uv_sphere
-from fast_trimesh.cpu.types import Affine3D
+from fast_trimesh import Affine3D, cuboid, icosphere, save_stl, uv_sphere
 
 num_iters = 4
