+++ src/cpu/ops.cpp
@@ -2,35 +2,38 @@
 
 #include <iostream>
 
+#include "three/types.h"
+#include "two/types.h"
+
 using namespace pybind11::literals;
 
 namespace fast_trimesh {
 namespace cpu {
 namespace ops {
 
-types::Trimesh3D linear_extrude(const types::Polygon2D &polygon, float height) {
+Trimesh3D linear_extrude(const Polygon2D &polygon, float height) {
     std::function<float(float, float)> height_func =
         [height](float x, float y) -> float { return height; };
     return linear_extrude(polygon, height_func);
 }
 
-types::Trimesh3D linear_extrude(
-    const types::Polygon2D &polygon,
+Trimesh3D linear_extrude(
+    const Polygon2D &polygon,
     const std::function<float(float, float)> &height_func) {
     return linear_extrude(polygon.get_trimesh(), height_func);
 }
 
-types::Trimesh3D linear_extrude(const types::Trimesh2D &polygon, float height) {
+Trimesh3D linear_extrude(const Trimesh2D &polygon, float height) {
     std::function<float(float, float)> height_func =
         [height](float x, float y) -> float { return height; };
     return linear_extrude(polygon, height_func);
 }
 
-types::Trimesh3D linear_extrude(
-    const types::Trimesh2D &mesh,
+Trimesh3D linear_extrude(
+    const Trimesh2D &mesh,
     const std::function<float(float, float)> &height_func) {
-    types::vertices3d_t vertices;
-    types::face_set_t faces;
+    std::vector<Point3D> vertices;
+    face_set_t faces;
 
     // Ensure that polygon is counter-clockwise.
     std::vector<size_t> poly_inds = mesh.get_polygon_inds();
@@ -78,8 +81,8 @@ types::Trimesh3D linear_extrude(
     return {vertices, faces};
 }
 
-types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
-                                int n, int axis) {
+Trimesh3D rotate_extrude(const Polygon2D &polygon, float angle, int n,
+                         int axis) {
     // Checks that `angle` is less than a full circle.
     if (angle < 0 || angle > 2 * M_PI - 1e-6) {
         throw std::invalid_argument(
@@ -97,11 +100,11 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
         throw std::invalid_argument("`axis` must be 0, 1, or 2.");
     }
 
-    types::vertices3d_t vertices;
-    types::face_list_t faces;
+    std::vector<Point3D> vertices;
+    face_list_t faces;
 
     // Ensure that polygon is counter-clockwise.
-    types::Polygon2D poly = polygon;
+    Polygon2D poly = polygon;
     if (poly.is_clockwise()) {
         poly.reverse();
     }
@@ -119,7 +122,7 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
         int v0 = i, v1 = i + 1, v2 = 0;
         // Note that the order of the vertices is reversed, because the
         // bottom is face-down.
-        faces.push_back(types::face_t(v0, v2, v1));
+        faces.push_back(face_t(v0, v2, v1));
     }
 
     // Adds side faces. Each side face is made up of two triangles.
@@ -129,11 +132,11 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
         std::tuple<float, float, float> rot{axis == 0 ? angle_i : 0.0,
                                             axis == 1 ? angle_i : 0.0f,
                                             axis == 2 ? angle_i : 0.0f};
-        types::Affine3D tf{rot};
+        Affine3D tf{rot};
 
         // Adds next layer of vertices.
         for (int j = 0; j < p; j++) {
-            types::Point3D v{poly.points[j].x, poly.points[j].y, 0.0f};
+            Point3D v{poly.points[j].x, poly.points[j].y, 0.0f};
             v <<= tf;
             vertices.push_back(v);
         }
@@ -142,8 +145,8 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
         for (int j = 0; j < p; j++) {
             int v0 = j + offset, v1 = (j + 1) % p + offset, v2 = j + offset + p,
                 v3 = (j + 1) % p + offset + p;
-            faces.push_back(types::face_t(v0, v1, v2));
-            faces.push_back(types::face_t(v1, v3, v2));
+            faces.push_back(face_t(v0, v1, v2));
+            faces.push_back(face_t(v1, v3, v2));
         }
 
         offset += p;
@@ -152,15 +155,14 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, float angle,
     // Adds top face.
     for (int i = 1; i < p - 1; i++) {
         int v0 = i + offset, v1 = i + 1 + offset, v2 = offset;
-        faces.push_back(types::face_t(v0, v1, v2));
+        faces.push_back(face_t(v0, v1, v2));
     }
 
-    types::Trimesh3D mesh3d = {vertices, faces};
+    Trimesh3D mesh3d = {vertices, faces};
     return mesh3d.signed_volume() < 0 ? mesh3d.flip_inside_out() : mesh3d;
 }
 
-types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, int n,
-                                int axis) {
+Trimesh3D rotate_extrude(const Polygon2D &polygon, int n, int axis) {
     // Checks that `n` is valid.
     if (n < 1) {
         throw std::invalid_argument("`n` must be at least 1.");
@@ -171,11 +173,11 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, int n,
         throw std::invalid_argument("`axis` must be 0, 1, or 2.");
     }
 
-    types::vertices3d_t vertices;
-    types::face_list_t faces;
+    std::vector<Point3D> vertices;
+    face_list_t faces;
 
     // Ensure that polygon is counter-clockwise.
-    types::Polygon2D poly = polygon;
+    Polygon2D poly = polygon;
     if (poly.is_clockwise()) {
         poly.reverse();
     }
@@ -197,12 +199,12 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, int n,
         std::tuple<float, float, float> rot{axis == 0 ? angle_i : 0.0,
                                             axis == 1 ? angle_i : 0.0f,
                                             axis == 2 ? angle_i : 0.0f};
-        types::Affine3D tf{rot};
+        Affine3D tf{rot};
         int offset_next = offset + p;
 
         // Adds next layer of vertices.
         for (int j = 0; j < p; j++) {
-            types::Point3D v{poly.points[j].x, poly.points[j].y, 0.0f};
+            Point3D v{poly.points[j].x, poly.points[j].y, 0.0f};
             v <<= tf;
             vertices.push_back(v);
         }
@@ -211,8 +213,8 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, int n,
         for (int j = 0; j < p; j++) {
             int v0 = j + offset, v1 = (j + 1) % p + offset,
                 v2 = j + offset_next, v3 = (j + 1) % p + offset_next;
-            faces.push_back(types::face_t(v0, v1, v2));
-            faces.push_back(types::face_t(v1, v3, v2));
+            faces.push_back(face_t(v0, v1, v2));
+            faces.push_back(face_t(v1, v3, v2));
         }
 
         offset = offset_next;
@@ -222,43 +224,39 @@ types::Trimesh3D rotate_extrude(const types::Polygon2D &polygon, int n,
     for (int j = 0; j < p; j++) {
         int v0 = j + offset, v1 = (j + 1) % p + offset, v2 = j,
             v3 = (j + 1) % p;
-        faces.push_back(types::face_t(v0, v1, v2));
-        faces.push_back(types::face_t(v1, v3, v2));
+        faces.push_back(face_t(v0, v1, v2));
+        faces.push_back(face_t(v1, v3, v2));
     }
 
-    types::Trimesh3D mesh3d{vertices, faces};
+    Trimesh3D mesh3d{vertices, faces};
     return mesh3d.signed_volume() < 0 ? mesh3d.flip_inside_out() : mesh3d;
 }
 
-void add_modules(py::module &m) {
-    py::module o = m.def_submodule("ops");
-    o.doc() = "Mesh operations.";
-
-    o.def("linear_extrude",
-          py::overload_cast<const types::Polygon2D &, float>(&linear_extrude),
+void add_ops_modules(py::module &m) {
+    m.def("linear_extrude",
+          py::overload_cast<const Polygon2D &, float>(&linear_extrude),
           "Linearly extrudes a 2D mesh", "mesh"_a, "height"_a);
-    o.def("linear_extrude",
-          py::overload_cast<const types::Polygon2D &,
+    m.def("linear_extrude",
+          py::overload_cast<const Polygon2D &,
                             const std::function<float(float, float)> &>(
               &linear_extrude),
           "Linearly extrudes a 2D mesh", "mesh"_a, "height_func"_a);
-    o.def("linear_extrude",
-          py::overload_cast<const types::Trimesh2D &, float>(&linear_extrude),
+    m.def("linear_extrude",
+          py::overload_cast<const Trimesh2D &, float>(&linear_extrude),
           "Linearly extrudes a 2D mesh", "mesh"_a, "height"_a);
-    o.def("linear_extrude",
-          py::overload_cast<const types::Trimesh2D &,
+    m.def("linear_extrude",
+          py::overload_cast<const Trimesh2D &,
                             const std::function<float(float, float)> &>(
               &linear_extrude),
           "Linearly extrudes a 2D mesh", "mesh"_a, "height_func"_a);
 
-    o.def("rotate_extrude",
-          py::overload_cast<const types::Polygon2D &, float, int, int>(
-              &rotate_extrude),
-          "Rotates a 2D mesh", "mesh"_a, "angle"_a, "n"_a, "axis"_a = 0);
-    o.def(
+    m.def(
         "rotate_extrude",
-        py::overload_cast<const types::Polygon2D &, int, int>(&rotate_extrude),
-        "Rotates a 2D mesh", "mesh"_a, "n"_a, "axis"_a = 0);
+        py::overload_cast<const Polygon2D &, float, int, int>(&rotate_extrude),
+        "Rotates a 2D mesh", "mesh"_a, "angle"_a, "n"_a, "axis"_a = 0);
+    m.def("rotate_extrude",
+          py::overload_cast<const Polygon2D &, int, int>(&rotate_extrude),
+          "Rotates a 2D mesh", "mesh"_a, "n"_a, "axis"_a = 0);
 }
 
 }  // namespace ops
