+++ src/cpu/shapes.h
@@ -3,6 +3,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
+#include "three/types.h"
+#include "two/types.h"
 #include "types.h"
 
 namespace py = pybind11;
@@ -11,22 +13,25 @@ namespace fast_trimesh {
 namespace cpu {
 namespace shapes {
 
-types::Polygon2D rectangle(float width, float height, bool center = false);
+using namespace fast_trimesh::cpu::types;
+using namespace fast_trimesh::cpu::two::types;
+using namespace fast_trimesh::cpu::three::types;
 
-types::Polygon2D regular_polygon(float radius, int n);
+Polygon2D rectangle(float width, float height, bool center = false);
 
-types::Trimesh2D regular_polygon_mesh(float radius, int n);
+Polygon2D regular_polygon(float radius, int n);
 
-types::Trimesh3D cuboid(float width, float height, float depth,
-                        bool center = false);
+Trimesh2D regular_polygon_mesh(float radius, int n);
 
-types::Trimesh3D tetrahedron(float radius);
+Trimesh3D cuboid(float width, float height, float depth, bool center = false);
 
-types::Trimesh3D icosphere(float radius, int n);
+Trimesh3D tetrahedron(float radius);
 
-types::Trimesh3D uv_sphere(float radius, int n, int m);
+Trimesh3D icosphere(float radius, int n);
 
-void add_modules(py::module &m);
+Trimesh3D uv_sphere(float radius, int n, int m);
+
+void add_shapes_modules(py::module &m);
 
 }  // namespace shapes
 }  // namespace cpu
