+++ src/cpu/three/boolean.cpp
@@ -13,17 +13,17 @@ namespace cpu {
 namespace three {
 namespace boolean {
 
-std::vector<std::tuple<size_t, size_t>> get_edges(const types::face_t &face) {
+std::vector<std::tuple<size_t, size_t>> get_edges(const face_t &face) {
     auto &[a, b, c] = face;
     return {{a, b}, {b, c}, {c, a}};
 }
 
 enum boolean_op { UNION, INTERSECTION, DIFFERENCE };
 
-types::Trimesh3D triangulation(const types::Triangle3D &triangle,
-                               const std::vector<types::Point3D> &points) {
-    types::vertices3d_t vertices;
-    types::face_set_t faces;
+Trimesh3D triangulation(const Triangle3D &triangle,
+                        const std::vector<Point3D> &points) {
+    std::vector<Point3D> vertices;
+    face_set_t faces;
 
     // Checks that all points are inside the triangle.
     for (auto &point : points) {
@@ -38,7 +38,7 @@ types::Trimesh3D triangulation(const types::Triangle3D &triangle,
     for (auto &point : points) vertices.push_back(point);
 
     // Adds super triangle to mesh.
-    faces.insert(types::face_t(0, 1, 2));
+    faces.insert(face_t(0, 1, 2));
 
     for (size_t i = 0; i < points.size(); i++) {
         auto &point = points[i];
@@ -47,11 +47,11 @@ types::Trimesh3D triangulation(const types::Triangle3D &triangle,
         bool found_triangle = false;
         for (auto &face : faces) {
             auto &[a, b, c] = face;
-            types::Triangle3D triangle{vertices[a], vertices[b], vertices[c]};
+            Triangle3D triangle{vertices[a], vertices[b], vertices[c]};
             if (point.projects_to_triangle(triangle)) {
-                faces.insert(types::face_t(a, b, i + 3));
-                faces.insert(types::face_t(b, c, i + 3));
-                faces.insert(types::face_t(c, a, i + 3));
+                faces.insert(face_t(a, b, i + 3));
+                faces.insert(face_t(b, c, i + 3));
+                faces.insert(face_t(c, a, i + 3));
                 faces.erase(face);
                 found_triangle = true;
                 break;
@@ -67,12 +67,11 @@ types::Trimesh3D triangulation(const types::Triangle3D &triangle,
     return {vertices, faces};
 }
 
-types::Trimesh3D mesh_union_no_intersections(const types::Trimesh3D &a,
-                                             const types::Trimesh3D &b) {
+Trimesh3D mesh_union_no_intersections(const Trimesh3D &a, const Trimesh3D &b) {
     const size_t n = a.vertices().size();
 
-    types::vertices3d_t vertices;
-    types::face_set_t faces;
+    std::vector<Point3D> vertices;
+    face_set_t faces;
 
     // Adds all vertices to the output mesh.
     vertices.insert(vertices.end(), a.vertices().begin(), a.vertices().end());
@@ -81,30 +80,30 @@ types::Trimesh3D mesh_union_no_intersections(const types::Trimesh3D &a,
     // Adds first mesh's faces to the output mesh.
     for (auto &face : a.faces()) {
         auto &[a, b, c] = face;
-        faces.insert(types::face_t(a, b, c));
+        faces.insert(face_t(a, b, c));
     }
 
     // Adds second mesh's faces to the output mesh, offset by n.
     for (auto &face : b.faces()) {
         auto &[a, b, c] = face;
-        faces.insert(types::face_t(a + n, b + n, c + n));
+        faces.insert(face_t(a + n, b + n, c + n));
     }
 
     return {vertices, faces};
 }
 
-bool is_inside(const types::Triangle3D &triangle, const types::Point3D &face_pt,
-               const types::Point3D &pt) {
-    types::Point3D normal = triangle.normal();
-    types::Point3D vec = pt - face_pt;
+bool is_inside(const Triangle3D &triangle, const Point3D &face_pt,
+               const Point3D &pt) {
+    Point3D normal = triangle.normal();
+    Point3D vec = pt - face_pt;
     float dp = normal.dot(vec);
     return dp < 0;
 }
 
-std::tuple<types::Trimesh3D, types::Trimesh3D> split_at_all_intersections(
-    const types::Trimesh3D &a, const types::Trimesh3D &b) {
-    types::vertices3d_t a_vertices, b_vertices;
-    types::face_set_t a_faces, b_faces;
+std::tuple<Trimesh3D, Trimesh3D> split_at_all_intersections(
+    const Trimesh3D &a, const Trimesh3D &b) {
+    std::vector<Point3D> a_vertices, b_vertices;
+    face_set_t a_faces, b_faces;
 
     // Adds all vertices to the output mesh.
     a_vertices.insert(a_vertices.end(), a.vertices().begin(),
@@ -113,7 +112,7 @@ std::tuple<types::Trimesh3D, types::Trimesh3D> split_at_all_intersections(
                       b.vertices().end());
 
     // Gets BVH for each mesh.
-    bvh::BVH a_bvh(a), b_bvh(b);
+    bvh::BVH3D a_bvh(a), b_bvh(b);
 
     // Stores IDs of vertices which are shared between meshes.
     std::vector<std::tuple<size_t, size_t>> vertex_pairs;
@@ -121,37 +120,30 @@ std::tuple<types::Trimesh3D, types::Trimesh3D> split_at_all_intersections(
     return {{a_vertices, a_faces}, {b_vertices, b_faces}};
 }
 
-types::Trimesh3D mesh_op(const types::Trimesh3D &a, const types::Trimesh3D &b,
-                         boolean_op op) {
+Trimesh3D mesh_op(const Trimesh3D &a, const Trimesh3D &b, boolean_op op) {
     throw std::runtime_error("Not implemented yet.");
 }
 
-types::Trimesh3D mesh_union(const types::Trimesh3D &a,
-                            const types::Trimesh3D &b) {
+Trimesh3D mesh_union(const Trimesh3D &a, const Trimesh3D &b) {
     return mesh_op(a, b, UNION);
 }
 
-types::Trimesh3D mesh_intersection(const types::Trimesh3D &a,
-                                   const types::Trimesh3D &b) {
+Trimesh3D mesh_intersection(const Trimesh3D &a, const Trimesh3D &b) {
     return mesh_op(a, b, INTERSECTION);
 }
 
-types::Trimesh3D mesh_difference(const types::Trimesh3D &a,
-                                 const types::Trimesh3D &b) {
+Trimesh3D mesh_difference(const Trimesh3D &a, const Trimesh3D &b) {
     return mesh_op(a, b, DIFFERENCE);
 }
 
-void add_modules(py::module &m) {
-    py::module s = m.def_submodule("boolean");
-    s.doc() = "Boolean operations";
-
-    s.def("triangulation", &triangulation,
+void add_3d_boolean_modules(py::module &m) {
+    m.def("triangulation", &triangulation,
           "Gets a triangulation of a triangle and a set of points",
           "triangle"_a, "points"_a);
-    s.def("union", &mesh_union, "Union of two meshes", "a"_a, "b"_a);
-    s.def("intersection", &mesh_intersection, "Intersection of two meshes",
+    m.def("union", &mesh_union, "Union of two meshes", "a"_a, "b"_a);
+    m.def("intersection", &mesh_intersection, "Intersection of two meshes",
           "a"_a, "b"_a);
-    s.def("difference", &mesh_difference, "Difference of two meshes", "a"_a,
+    m.def("difference", &mesh_difference, "Difference of two meshes", "a"_a,
           "b"_a);
 }
