+++ src/cpu/three/bvh.h
@@ -6,6 +6,7 @@
 #include <unordered_map>
 
 #include "../types.h"
+#include "types.h"
 
 namespace py = pybind11;
 
@@ -14,36 +15,26 @@ namespace cpu {
 namespace three {
 namespace bvh {
 
-typedef std::tuple<size_t, size_t> edge_t;
-
-struct IntersectionSet {
-    std::vector<std::vector<size_t>> a_to_b;
-    std::vector<std::vector<size_t>> b_to_a;
-    std::vector<std::tuple<edge_t, types::Point3D>> points;
-
-    void add(const size_t a, const size_t b, const types::Point3D &p);
-    std::string to_string() const;
-};
-
-IntersectionSet intersections(const types::Trimesh3D &a,
-                              const types::Trimesh3D &b);
+using namespace fast_trimesh::cpu::types;
+using namespace fast_trimesh::cpu::three::types;
 
 typedef std::tuple<size_t, size_t> edge_t;
 
 struct __edge_hash_fn {
     std::size_t operator()(const edge_t &f) const {
-        return std::get<0>(f) ^ std::get<1>(f);
+        auto hf = std::hash<size_t>();
+        return hf(std::get<0>(f)) ^ hf(std::get<1>(f));
     }
 };
 
 struct TrimeshAdjacency {
     std::vector<std::vector<size_t>> vertex_to_faces;
     std::vector<std::vector<size_t>> vertex_to_vertices;
-    std::vector<types::face_t> face_to_vertices;
-    std::vector<types::face_t> face_to_faces;
+    std::vector<face_t> face_to_vertices;
+    std::vector<face_t> face_to_faces;
     std::unordered_map<edge_t, size_t, __edge_hash_fn> edge_to_faces;
 
-    TrimeshAdjacency(const types::Trimesh3D &mesh);
+    TrimeshAdjacency(const Trimesh3D &mesh);
 
     void validate() const;
 };
@@ -56,24 +47,24 @@ struct TrimeshAdjacency {
 // The triangle ID points to the triangle in the trimesh. If a ray
 // doesn't intersect a box, then we don't need to check any of the
 // triangles in the box.
-typedef std::vector<std::tuple<size_t, int, int, types::BoundingBox3D>> tree_t;
+typedef std::vector<std::tuple<size_t, int, int, BoundingBox3D>> tree_t;
 
-struct BVH {
-    const std::shared_ptr<types::Trimesh3D> trimesh;
+struct BVH3D {
+    const std::shared_ptr<Trimesh3D> trimesh;
     tree_t tree;
 
-    BVH(const types::Trimesh3D &t);
-    ~BVH() = default;
-    const std::shared_ptr<types::Trimesh3D> get_trimesh() const {
+    BVH3D(const Trimesh3D &t);
+    ~BVH3D() = default;
+    const std::shared_ptr<Trimesh3D> get_trimesh() const {
         return this->trimesh;
     }
     tree_t get_tree() const { return this->tree; }
-    std::vector<std::tuple<size_t, types::face_t, types::Point3D>>
-    intersections(const types::Line3D &l) const;
+    std::vector<std::tuple<size_t, face_t, Point3D>> intersections(
+        const Line3D &l) const;
     std::string to_string() const;
 };
 
-void add_modules(py::module &m);
+void add_3d_bvh_modules(py::module &m);
 
 }  // namespace bvh
 }  // namespace three
