+++ src/cpu/three/main.h
@@ -5,16 +5,13 @@
 
 #include <vector>
 
-#include "boolean.h"
-#include "bvh.h"
-
 namespace py = pybind11;
 
 namespace fast_trimesh {
 namespace cpu {
 namespace three {
 
-void add_modules(py::module &m);
+void add_3d_modules(py::module &m);
 
 }
 }  // namespace cpu
