+++ src/cpu/two/boolean.h
@@ -4,6 +4,7 @@
 #include <pybind11/stl.h>
 
 #include "../types.h"
+#include "types.h"
 
 namespace py = pybind11;
 
@@ -12,24 +13,23 @@ namespace cpu {
 namespace two {
 namespace boolean {
 
+using namespace fast_trimesh::cpu::types;
+using namespace fast_trimesh::cpu::two::types;
+
 // Given a triangle and a set of points, return a trimesh that is the
 // triangulation of the triangle with respect to the points.
-types::Trimesh2D triangulation(const types::Triangle2D &triangle,
-                               const std::vector<types::Point2D> &points);
+Trimesh2D triangulation(const Triangle2D &triangle,
+                        const std::vector<Point2D> &points);
 
 // Given two trimeshes, splits the trimeshes at all intersections between
 // them. Returns a pair of trimeshes, where
-types::Trimesh2D split_at_all_intersections(const types::Trimesh2D &a,
-                                            const types::Trimesh2D &b);
+Trimesh2D split_at_all_intersections(const Trimesh2D &a, const Trimesh2D &b);
 
-types::Trimesh2D mesh_union(const types::Trimesh2D &a,
-                            const types::Trimesh2D &b);
-types::Trimesh2D mesh_intersection(const types::Trimesh2D &a,
-                                   const types::Trimesh2D &b);
-types::Trimesh2D mesh_difference(const types::Trimesh2D &a,
-                                 const types::Trimesh2D &b);
+Trimesh2D mesh_union(const Trimesh2D &a, const Trimesh2D &b);
+Trimesh2D mesh_intersection(const Trimesh2D &a, const Trimesh2D &b);
+Trimesh2D mesh_difference(const Trimesh2D &a, const Trimesh2D &b);
 
-void add_modules(py::module &m);
+void add_2d_boolean_modules(py::module &m);
 
 }  // namespace boolean
 }  // namespace two
