+++ src/cuda/main.h
@@ -7,7 +7,7 @@ namespace py = pybind11;
 namespace fast_trimesh {
 namespace cuda {
 
-void add_modules(py::module &m);
+void add_cuda_modules(py::module &m);
 
 }  // namespace cuda
 }  // namespace fast_trimesh
