+++ tests/test_cpu_trimesh.py
@@ -5,7 +5,11 @@ from pathlib import Path
 
 import pytest
 
-from fast_trimesh.cpu.io import (
+from fast_trimesh import (
+    Affine3D,
+    Point2D,
+    Polygon2D,
+    cuboid,
     load_obj,
     load_ply,
     load_stl,
@@ -15,8 +19,6 @@ from fast_trimesh.cpu.io import (
     save_stl,
     save_stl_text,
 )
-from fast_trimesh.cpu.shapes import cuboid
-from fast_trimesh.cpu.types import Affine3D, Point2D, Polygon2D, Trimesh2D
 
 
 @pytest.mark.parametrize(
